Refactor array indexing and plotting routines
Updated multiple files to remove unnecessary float conversions and use direct array indexing for improved clarity and consistency. Replaced deprecated np.trapz with np.trapezoid in rotor noise calculations. Refactored rotor performance plotting to use matplotlib instead of plotly, and standardized legend and axis handling. Adjusted weight analysis result files to correct payload and operational item calculations.

+++ RCAIDE/Library/Methods/Powertrain/Sources/Fuel_Tanks/Integral_Tank/compute_integral_tank_volume.py
@@ -125,7 +125,7 @@ def compute_fuselage_integral_tank_fuel_volume(fuel_tank,fuselage):
             if actual_fuel_volume > fuel_tank.volume_properties.net_volume + 1e-8 :
                 raise AttributeError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
         else:
-            fuel_tank.fuel.mass_properties.mass = float(tank_volume_i *  fuel_tank.fuel.density)    
+            fuel_tank.fuel.mass_properties.mass           = tank_volume_i *  fuel_tank.fuel.density    
             fuel_tank.fuel.volume_properties.gross_volume = tank_volume_i
             
     # update orign of tank 
@@ -257,7 +257,7 @@ def compute_wing_integral_tank_volume(fuel_tank,wing):
             if actual_fuel_volume > fuel_tank.volume_properties.net_volume + 1e-8 :
                 raise AttributeError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
         else:
-            fuel_tank.fuel.mass_properties.mass = float(total_fuel_volume *  fuel_tank.fuel.density)  
+            fuel_tank.fuel.mass_properties.mass = total_fuel_volume *  fuel_tank.fuel.density   
             
     else:  
         # assume whole wing has fuel 
@@ -275,7 +275,7 @@ def compute_wing_integral_tank_volume(fuel_tank,wing):
             raise AttributeError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
         fuel_tank.fuel.volume_properties.net_volume = actual_fuel_volume
     else:
-        fuel_tank.fuel.mass_properties.mass = float(total_fuel_volume *  fuel_tank.fuel.density)  
+        fuel_tank.fuel.mass_properties.mass         = total_fuel_volume *  fuel_tank.fuel.density   
         fuel_tank.fuel.volume_properties.net_volume = total_fuel_volume    
     return
