+++ RCAIDE/Library/Methods/Powertrain/Sources/Fuel_Tanks/Non_Integral_Tank/compute_non_integral_tank_volume.py
@@ -236,7 +236,7 @@ def compute_bwb_aft_tank_volume(fuel_tank, wing):
             raise ValueError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank')
         fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
     else:
-        fuel_tank.fuel.mass_properties.mass = float(tank_volume_i *  fuel_tank.fuel.density)
+        fuel_tank.fuel.mass_properties.mass         = tank_volume_i *  fuel_tank.fuel.density
         fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
     # fuel tank origin
     fuel_tank.origin[0][0]  = circle_origins[max_idx][0] - fuel_tank.outer_diameter / 2
@@ -295,7 +295,7 @@ def compute_prismatic_fuel_tank_volume(fuel_tank):
         if actual_fuel_volume > fuel_tank.volume_properties.net_volume + 1e-8 :
             raise ValueError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
     else:
-        fuel_tank.fuel.mass_properties.mass     = float(tank_volume_i *  fuel_tank.fuel.density)
+        fuel_tank.fuel.mass_properties.mass         = tank_volume_i *  fuel_tank.fuel.density 
         fuel_tank.fuel.volume_properties.net_volume = tank_volume_i 
     
     fuel_tank.fuel.mass_properties.center_of_gravity  =  [[fuel_tank.outer_length /2, 0, 0]] 
@@ -376,7 +376,7 @@ def compute_wing_non_integral_tank_volume(fuel_tank, wing,fuel_tanks):
                     raise AttributeError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
                 fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
             else:
-                fuel_tank.fuel.mass_properties.mass = float(tank_volume_i *  fuel_tank.fuel.density)
+                fuel_tank.fuel.mass_properties.mass = tank_volume_i *  fuel_tank.fuel.density 
                 fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
              
     return
